requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
celery==5.4.0
redis==5.0.8
openai==1.51.2
praw==7.7.1
httpx==0.27.2
requests==2.32.3
python-multipart==0.0.12
email-validator==2.2.0
jinja2==3.1.4
gunicorn==23.0.0
uvloop==0.20.0
httptools==0.6.1
orjson==3.10.7
sentry-sdk==2.16.0
textstat==0.7.3
websockets==12.0
psutil==5.9.8

